+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Increase route highlight thickness.
 - Automatically pause uninitialized editor.
+- Remove "Deselect filtered" button.
+- Improve hold picker look and behavior on smaller screens.
 - Fix preferences being inaccessible behind the pause sceen.
 - Fix raycasting being applied to starting hold markers.